+++ components/ColourPicker/libs/colour_models/public/colour_models.hpp
@@ -2,7 +2,7 @@
 
 namespace colour_picker{
     void HSLToRGB(
-        const float hue, const float saturation, const float lightness,
+        const double hue, const double saturation, const double lightness,
         uint8_t& red, uint8_t& green, uint8_t& blue
     );
 }
